+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## DEV
+
+* Improved handling JSON parsing error
+
 ## 0.1.0 (2019-02-15)
 
 * First release